[CodeHealth][Android] Move BraveConstants out of deprecated `util` package
Move BraveConstants.java from org.chromium.chrome.browser.util to
org.chromium.chrome.browser to align with upstream's hierarchy.
Upstream's analogous ChromeStringConstants lives directly under
org.chromium.chrome.browser, and the util package is treated as
deprecated for new top-level constants/utilities.

No behavior changes — pure refactor.

Resolves: https://github.com/brave/brave-browser/issues/55685
