+++ android/java/org/chromium/chrome/browser/BraveConstants.java
@@ -1,12 +1,13 @@
-/**
- * Copyright (c) 2022 The Brave Authors. All rights reserved.
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
 
-package org.chromium.chrome.browser.util;
+package org.chromium.chrome.browser;
 
+import org.chromium.build.annotations.NullMarked;
+
+@NullMarked
 public final class BraveConstants {
     // used in BraveRewardsPanel
     public static final int SITE_BANNER_REQUEST_CODE = 33;
